split up the components
